Revert "Clarified output."
This reverts commit 091fc5bc3b8439cc863460501f8b3ebf45a96b67.
The changed that were introduced, worked worse than anticipated.
The solution, is instead to make sure that all files are saved
without spaces in their names. This requires a mod to
downloadAssignment, and in uploadFeedback it should abort if a
filename with space is found.
A helper (fixfilenames.sh) will be added to handle legacy
submissions.