Added rudimentary suppor to handle group submissions. It 'relies' on an experimental beta feature.
If a submission is done as a group, the save folder will match the
group name. In the folder a file called GROUP.txt will be found.
Each row contains the info of a user in that group.

Note, that uploadFeedback needs to be updated as to also support this.
ATM, it will probably NOT work. But adaptation should be easy,
as the feedback should just add an option 'group', to indicate
group feedback.